+++ app/src/main/java/org/wso2/oidc/sample/LoginActivity.java
@@ -1,4 +1,22 @@
-package com.example.myapplication;
+/*
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.oidc.sample;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.browser.customtabs.CustomTabsIntent;
@@ -8,7 +26,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 
-
+import com.example.sample.R;
 import org.oidc.agent.sso.LoginService;
 
 import java.util.concurrent.atomic.AtomicReference;
