--- app/src/main/res/raw/wso2carbon.pem
@@ -1,21 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDdzCCAl+gAwIBAgIJAOwBeR+cuxEDMA0GCSqGSIb3DQEBCwUAMGExCzAJBgNV
-BAYTAlNMMRAwDgYDVQQIEwd3ZXN0ZXJuMRAwDgYDVQQHEwdjb2xvbWJvMQswCQYD
-VQQKEwJpczENMAsGA1UECxMEd3NvMjESMBAGA1UEAxMJbG9jYWxob3N0MB4XDTIw
-MDUxODEwMDY0NFoXDTIwMDgxNjEwMDY0NFowYTELMAkGA1UEBhMCU0wxEDAOBgNV
-BAgTB3dlc3Rlcm4xEDAOBgNVBAcTB2NvbG9tYm8xCzAJBgNVBAoTAmlzMQ0wCwYD
-VQQLEwR3c28yMRIwEAYDVQQDEwlsb2NhbGhvc3QwggEiMA0GCSqGSIb3DQEBAQUA
-A4IBDwAwggEKAoIBAQCnRPGcIroGe8ZUcUbvV42KqvwrVSz4MoR57M4iVQgzOCOh
-H93aNFW1HUxTjmBNKGanS2o1+jZir8BD+g6lm53lon9ocG6xE+O7taUYzmwV+9eM
-RjTlzhHjU0m2GFHCAFympndSiSqPi8hBoN/46KEXX5ww3LgblvvHWxMvq2VkqzTZ
-Sc1gSKMY0T2od/yUx67DQs4wKUOJH6UVKw7n7A9EY4L1JUnagpW6+a+xUIhcImJp
-gWhfgEzp3hvrYVKSEezsVNqRnJmEp+vm0pBW0n/dtGgjQKqRJO3xEvG7rb0Y9zA5
-0WBSfXhuTdY69gewCwf0dkjWS7kg9MVnbdaXf5NBAgMBAAGjMjAwMB0GA1UdDgQW
-BBSvYSva4tYGl/xK1waT/EjsBwKoRTAPBgNVHREECDAGhwQKAAICMA0GCSqGSIb3
-DQEBCwUAA4IBAQCeZmWLR6LqMCI4H4o5MjPhyLnnGESKnbRQuC9yNyjIWUawZSDh
-71CRtKsffmi04Mz3eoVlt5bIbGof2jT7j/K4/xDrXlNsjz9Z6M1aRFF7zfWGFS43
-YT6as2Cl4SoPYcznKHG7lFTz4+IfBQKvddVM/WHdf4jjgnYrgj5Y0wWWUtnRuSYM
-31jVXdjuXtvt+0F9gY3ZzM1MZoVqP0l/kT03I/UKQldDC5DAd9DXlQBlDIdGtO7H
-OBfcJw4CRnY1XppNLQlQCGU0hasLhhmjO6qQeehoVQgNMCGFRsrYGErLp2nmLebA
-w/Q1lfboZwKJP+XFZR+oGAFNK6DaokjTdF0a
------END CERTIFICATE-----
