--- app/src/main/res/xml/network_security_config.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<network-security-config>
-    <domain-config>
-        <domain includeSubdomains="true">10.0.2.2</domain>
-        <trust-anchors>
-            <certificates src="@raw/wso2carbon"/>
-            <certificates src="@raw/wso2carbon1"/>
-        </trust-anchors>
-    </domain-config>
-</network-security-config>
