+++ build.gradle
@@ -1,12 +1,29 @@
+/*
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
     
     repositories {
-        //mavenLocal()
+        mavenLocal()
         google()
         jcenter()
-        maven{url "http://192.168.8.100:8081/nexus/service/local/staging/deploy/maven2"}
 
 
     }
@@ -22,10 +39,10 @@ buildscript {
 
 allprojects {
     repositories {
-        //mavenLocal()
+        mavenLocal()
         google()
         jcenter()
-        maven{url "http://192.168.8.100:8081/nexus/content/repositories/releases"}
+        maven{url "http://maven.wso2.org/nexus/content/repositories/releases"}
         
     }
 }
